NAI-73 Update naming convention and fix model relationships

+++ migrations/versions/0afdb03f9fea_create_users_table.py
@@ -18,7 +18,7 @@ depends_on = None
 
 def upgrade() -> None:
     op.create_table('users',
-        sa.Column('id', sa.Integer(), nullable=False),
+        sa.Column('id', sa.BigInteger(), nullable=False),
         sa.Column('email', sa.String(), nullable=False),
         sa.Column('password', sa.String(), nullable=False),
         sa.PrimaryKeyConstraint('id'),
